final/scripts/generate-shop/requirements.txt
python-dotenv~=1.2.1